tensorflow_datasets/datasets/beom_dataset/.ipynb_checkpoints/beam_requirements-checkpoint.txt
tensorflow_datasets[beom_dataset]